requirements.txt
openpyxl>=3.1.2
pandas>=2.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
loguru>=0.7.2
crewai>=0.1.0
streamlit>=1.28.0
plotly>=5.17.0

